Fix error compiling runscript code on 64 bit machines.
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@6673 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/dird/dird_conf.c
@@ -1945,7 +1945,7 @@ static void store_runscript(LEX *lc, RES_ITEM *item, int index, int pass)
        */
       res_runscript.set_job_code_callback(job_code_callback_filesetname);
       while ((c=(char*)res_runscript.commands->pop()) != NULL) {
-         t = (int) res_runscript.commands->pop();
+         t = (long)res_runscript.commands->pop();
          RUNSCRIPT *script = new_runscript();
          memcpy(script, &res_runscript, sizeof(RUNSCRIPT));
          script->command = c;
